requirements.txt
asgiref==3.5.2
awsebcli==3.20.3
botocore==1.23.54
cement==2.8.2
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.3
distlib==0.3.6
Django==4.1.1
django-crispy-forms==1.14.0
djangorestframework==3.13.1
filelock==3.8.0
future==0.16.0
idna==3.3
jmespath==0.10.0
pathspec==0.9.0
platformdirs==2.5.2
PyJWT==2.4.0
pypiwin32==223
python-dateutil==2.8.2
pytz==2022.2.1
pywin32==304
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
twilio==7.13.0
tzdata==2022.2
urllib3==1.26.12
virtualenv==20.16.5
virtualenvwrapper-win==1.2.7
wcwidth==0.1.9
